+++ core-ng/src/main/java/core/framework/internal/db/DatabaseImpl.java
@@ -7,6 +7,7 @@ import core.framework.db.IsolationLevel;
 import core.framework.db.Repository;
 import core.framework.db.Transaction;
 import core.framework.db.UncheckedSQLException;
+import core.framework.internal.db.cloud.CloudAuthProvider;
 import core.framework.internal.db.cloud.GCloudAuthProvider;
 import core.framework.internal.log.ActionLog;
 import core.framework.internal.log.LogLevel;
@@ -38,6 +39,8 @@ import static core.framework.log.Markers.errorCode;
  * @author neo
  */
 public final class DatabaseImpl implements Database {
+    public static final String PROPERTY_KEY_AUTH_PROVIDER = "authProvider";
+
     static {
         // disable unnecessary mysql connection cleanup thread to reduce overhead
         System.setProperty(PropertyDefinitions.SYSP_disableAbandonedConnectionCleanup, "true");
@@ -58,7 +61,7 @@ public final class DatabaseImpl implements Database {
 
     private String url;
     private Properties driverProperties;
-    private GCloudAuthProvider authProvider;     // generalize if support more clouds later
+    private CloudAuthProvider authProvider;
     private Duration timeout;
     private Driver driver;
 
@@ -93,7 +96,11 @@ public final class DatabaseImpl implements Database {
             driverProperties = driverProperties(url);
             this.driverProperties = driverProperties;
         }
-        if (authProvider != null) authProvider.fetchCredential(driverProperties);
+        if (authProvider != null) {
+            // properties are thread safe, it's ok to set user/password with multiple threads
+            driverProperties.setProperty("user", authProvider.user());
+            driverProperties.setProperty("password", authProvider.accessToken());
+        }
         Connection connection = null;
         try {
             connection = driver.connect(url, driverProperties);
@@ -134,6 +141,7 @@ public final class DatabaseImpl implements Database {
             // refer to https://cloud.google.com/sql/docs/mysql/authentication
             if (authProvider != null) {
                 properties.setProperty(PropertyKey.sslMode.getKeyName(), PropertyDefinitions.SslMode.PREFERRED.name());
+                properties.setProperty(PROPERTY_KEY_AUTH_PROVIDER, "gcloud");
             } else if (index == -1 || url.indexOf("sslMode=", index + 1) == -1) {
                 properties.setProperty(PropertyKey.sslMode.getKeyName(), PropertyDefinitions.SslMode.DISABLED.name());
             }
@@ -157,7 +165,11 @@ public final class DatabaseImpl implements Database {
 
     public void authProvider(String provider) {
         logger.info("use cloud auth provider, provider={}", provider);
-        authProvider = new GCloudAuthProvider();
+        if ("gcloud".equals(provider)) {
+            authProvider = GCloudAuthProvider.INSTANCE;
+        } else {
+            throw new Error("unsupported cloud auth provider, provider=" + provider);
+        }
     }
 
     public void url(String url) {