+++ core-ng/src/main/java/core/framework/internal/db/cloud/CloudAuthProvider.java
@@ -1,0 +1,12 @@
+package core.framework.internal.db.cloud;
+
+/**
+ * @author neo
+ */
+public interface CloudAuthProvider {
+
+
+    String user();
+
+    String accessToken();
+}