+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 7.10.6-b0 (04/19/2022 - )
 
-* db: tweak gcloud iam auth provider expiration time
+* db: tweak gcloud iam auth provider expiration time, make CancelQueryTaskImpl be aware of gcloud auth provider
 * db: tweak query timeout handling
   > increase db socket timeout, to make MySQL cancel timer (CancelQueryTaskImpl) has more room to kill query
 * http: update undertow to 2.2.17